wait on action to be completed
Wait for an applied action to be completed before returning. Not sure whether we should keep this like this as it is a blocking call. 
This was done because some actions in the climate entity can take several seconds to complete before the new state should be retrieved for an update. This way avoids adding a sleep.